Cambios en el main, escribe y lee, faltan correcciones en el índice

+++ book.cpp
@@ -52,7 +52,7 @@ bool Book::isMarked(){
 void Book::setMarked(bool marked){
 	this->marked = marked;
 }
-int Book::getMark(){
+long int Book::getMark(){
 	if(marked)
 		return mark;
 }
